refactor(아이콘): 아이콘 png -> svg 변경 (#12) - svg 아이콘 컴포넌트처럼 사용하기 위해 plugin 설치 및 설정 - 아이콘 파일 전부 png -> svg 교체
