+++ src/layouts/BottomTabBar.tsx
@@ -1,13 +1,11 @@
 import { Link } from "react-router-dom";
-import HomeIcon from "../components/icons/bottom-tab-menu/bottom-home.png";
-import CommunityIcon from "../components/icons/bottom-tab-menu/bottom-community.png";
-import NotificationIcon from "../components/icons/bottom-tab-menu/bottom-notify.png";
-import ProfileIcon from "../components/icons/bottom-tab-menu/bottom-profile.png";
-import HomeActiveIcon from "../components/icons/bottom-tab-menu/bottom-home-active.png";
-import CommunityActiveIcon from "../components/icons/bottom-tab-menu/bottom-community-active.png";
-import NotificationActiveIcon from "../components/icons/bottom-tab-menu/bottom-notify-active.png";
-import ProfileActiveIcon from "../components/icons/bottom-tab-menu/bottom-profile-active.png";
 import Text from "../components/common/Text";
+
+// icons
+import HomeIcon from  '../components/icons/bottom-tab-menu/bottom-home.svg?react';
+import CommunityIcon from  '../components/icons/bottom-tab-menu/bottom-community.svg?react';
+import NotifyIcon from  '../components/icons/bottom-tab-menu/bottom-notify.svg?react';
+import ProfileIcon from  '../components/icons/bottom-tab-menu/bottom-profile.svg?react';
 
 /**
  * 바텀 탭 메뉴 컴포넌트 인터페이스
@@ -21,15 +19,13 @@ interface IBottomTabBarProps {
 
 /**
  * 바텀 탭 메뉴 항목 인터페이스
- * @param iconSrc 아이콘 이미지 경로
- * @param iconSrcActive 활성화 아이콘 이미지 경로
+ * @param icon 아이콘 컴포넌트
  * @param text 탭 바 메뉴 텍스트
  * @param href 이동 경로
  * @param isActive 활성화 여부
  */
 interface IBottomTabItem {
-  iconSrc: string
-  iconSrcActive: string
+  icon: React.FunctionComponent<React.SVGProps<SVGSVGElement>>
   text: string
   href?: string
   isActive?: boolean
@@ -37,35 +33,36 @@ interface IBottomTabItem {
 
 /**
  * 바텀 탭 메뉴 항목 컴포넌트
- * @param iconSrc 아이콘 이미지 경로
+ * @param icon 아이콘 컴포넌트
  * @param text 탭 바 메뉴 텍스트
  * @param href 이동 경로
  * @param isActive 활성화 여부
  * @returns 탭 메뉴 아이템
  */
-function BottomTabBarItem({ iconSrc, iconSrcActive, text, href, isActive = false }: IBottomTabItem) {
-  const iconClassName = `h-5 w-5 ${isActive ? "opacity-100" : "opacity-90"}`
+function BottomTabBarItem({ icon, text, href, isActive = false }: IBottomTabItem) {
+  const iconClassName = `h-5 w-5 ${isActive ? "opacity-100 text-primary" : "opacity-90"}`;
+  const Icon = icon;
 
   const content = (
     <>
-      <img src={isActive ? iconSrcActive : iconSrc} alt={text} className={iconClassName} />
+      <Icon className={iconClassName} />
       <Text variant="REGULAR_12">{text}</Text>
     </>
-  )
+  );
 
   return (
     <li className="flex flex-col items-center justify-center gap-1">
       {href ? <Link to={href} className="flex flex-col items-center justify-center gap-1">{content}</Link> : content}
     </li>
-  )
+  );
 }
 
 export default function BottomTabBar({ isHome, isCommunity }: IBottomTabBarProps) {
   const bottomTabBarItems: IBottomTabItem[] = [
-    { iconSrc: HomeIcon, iconSrcActive: HomeActiveIcon, text: "홈", href: "/", isActive: isHome },
-    { iconSrc: CommunityIcon, iconSrcActive: CommunityActiveIcon, text: "게시글", href: "/community", isActive: isCommunity },
-    { iconSrc: NotificationIcon, iconSrcActive: NotificationActiveIcon, text: "알림" },
-    { iconSrc: ProfileIcon, iconSrcActive: ProfileActiveIcon, text: "마이" },
+    { icon: HomeIcon, text: "홈", href: "/", isActive: isHome },
+    { icon: CommunityIcon, text: "게시글", href: "/community", isActive: isCommunity },
+    { icon: NotifyIcon, text: "알림" },
+    { icon: ProfileIcon, text: "마이" },
   ]
 
   return (
@@ -74,8 +71,7 @@ export default function BottomTabBar({ isHome, isCommunity }: IBottomTabBarProps
         {bottomTabBarItems.map((item) => (
           <BottomTabBarItem
             key={item.text}
-            iconSrc={item.iconSrc}
-            iconSrcActive={item.iconSrcActive}
+            icon={item.icon}
             text={item.text}
             href={item.href}
             isActive={item.isActive}
