+++ src/layouts/MobileHeader.tsx
@@ -1,4 +1,4 @@
-import HeaderIcon from "../components/icons/header/header.png";
+import HeaderIcon from "../components/icons/header/header.svg?react";
 import Text from "../components/common/Text";
 
 interface IMobileHeaderProps {
@@ -9,7 +9,7 @@ export default function MobileHeader({ title }: IMobileHeaderProps) {
   return (
     <header className="shrink-0 bg-white px-4 pt-10">
       <div className="flex items-center justify-center gap-3">
-        <img src={HeaderIcon} alt="header" className="h-4 w-5" />
+        <HeaderIcon className="h-5 w-6" />
         <Text variant="SEMIBOLD_15">{title}</Text>
       </div>
     </header>
